requirements.txt
# Computer Vision & Detection
ultralytics>=8.2.0
opencv-python>=4.8.0.74
numpy>=1.26.0

# Facial Recognition & Emotion
deepface>=0.0.79
face-recognition>=1.3.0
dlib>=19.24.0
tf-keras>=2.15.0

# Data & Analytics
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# LLM Integration (Llama via Ollama)
requests>=2.31.0
# Ollama deve essere installato separatamente: https://ollama.ai

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Configuration & I/O
pyyaml>=6.0.1

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Additional utilities
python-dateutil>=2.8.2
tqdm>=4.66.0

# Database (built-in)
# sqlite3 is included with Python

# MLOps & Model Management
mlflow>=2.9.2
dvc>=3.50.0
onnx>=1.16.0
onnxruntime>=1.17.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-instrumentation-redis>=0.43b0
opentelemetry-instrumentation-asyncpg>=0.43b0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Event Bus
redis>=5.0.1

# S3-compatible storage (for MLflow artifacts)
boto3>=1.34.0

# API Gateway
httpx>=0.26.0
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
pylint>=3.0.0
mypy>=1.8.0
coverage-badge>=1.1.0

# Optional: per performance migliori
# torch>=2.1.0  # Per modelli PyTorch
